requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.4.0
transformers>=4.36.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
datasets>=2.14.0

# Async and HTTP
aiohttp>=3.8.0
httpx>=0.25.0
asyncio-throttle>=1.0.0

# Configuration and logging
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Evaluation and metrics
scikit-learn>=1.3.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

# Database
sqlite3
psycopg2-binary>=2.9.0

# Optional: GPU acceleration
# torch[cuda]>=2.0.0

# Web interface (optional)
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0